refactor to common module and app e2e pass
