+++ backend/package.json
@@ -28,6 +28,7 @@
     "@nestjs/platform-express": "^11.0.1",
     "@nestjs/typeorm": "^11.0.0",
     "bcryptjs": "^3.0.2",
+    "class-validator": "^0.14.2",
     "jsonwebtoken": "^9.0.2",
     "mysql2": "^3.14.1",
     "pg": "^8.16.0",
