+++ backend/src/common/decorators/role.decorator.ts
@@ -1,4 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
-import { UserRole } from '../../common/enum/user-role.enum';
+import { UserRole } from '../enum/user-role.enum';
 
 export const Role = (role: UserRole) => SetMetadata('role', role);
